refactor: 💡 environmentApply file
